refactor(spec): move recipe parsing into parser module
